Fixed update of coin image

+++ crypto_app_backend/database/migrations/2023_04_06_185717_create_transactions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('address');
             $table->string('username');
-            $table->float('amount');
+            $table->decimal('amount');
             $table->string('symbol');
             $table->string('type');
             $table->string('status');
